Renamed shampoo sales to generic product sales so that this sample will be suitable for all kinds of product sales data
