--- samples/csharp/end-to-end-apps/SpikeDetection-ShampooSales-WinForms/ShampooSalesConsoleApp/DataStructures/ShampooSalesData.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.ML.Data;
-
-namespace ShampooSalesAnomalyDetection
-{
-    public class ShampooSalesData
-    {
-        [LoadColumn(0)]
-        public string Month;
-
-        [LoadColumn(1)]
-        public float numSales;
-    }
-}
