--- samples/csharp/end-to-end-apps/SpikeDetection-ShampooSales-WinForms/ShampooSalesConsoleApp/DataStructures/ShampooSalesPrediction.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.ML.Data;
-
-namespace ShampooSalesAnomalyDetection
-{
-    class ShampooSalesPrediction
-    {
-        // Vector to hold Alert, Score, and P-Value values
-        [VectorType(3)]
-        public double[] Prediction { get; set; }
-    }
-}
