--- samples/csharp/end-to-end-apps/SpikeDetection-ShampooSales-WinForms/ShampooSalesSpikeDetection.sln
@@ -1,41 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.539
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ShampooSalesSpikesConsoleApp", "ShampooSalesConsoleApp\ShampooSalesSpikesConsoleApp.csproj", "{7845CC34-9B79-4834-9BE6-53ED7DFCA64A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShampooSalesSpikesWinForms", "ShampooSalesWinforms\ShampooSalesSpikesWinForms.csproj", "{8B051595-03CD-4026-91BB-403F90CF8526}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7845CC34-9B79-4834-9BE6-53ED7DFCA64A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7845CC34-9B79-4834-9BE6-53ED7DFCA64A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7845CC34-9B79-4834-9BE6-53ED7DFCA64A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{7845CC34-9B79-4834-9BE6-53ED7DFCA64A}.Debug|x64.Build.0 = Debug|Any CPU
-		{7845CC34-9B79-4834-9BE6-53ED7DFCA64A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7845CC34-9B79-4834-9BE6-53ED7DFCA64A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7845CC34-9B79-4834-9BE6-53ED7DFCA64A}.Release|x64.ActiveCfg = Release|Any CPU
-		{7845CC34-9B79-4834-9BE6-53ED7DFCA64A}.Release|x64.Build.0 = Release|Any CPU
-		{8B051595-03CD-4026-91BB-403F90CF8526}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8B051595-03CD-4026-91BB-403F90CF8526}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8B051595-03CD-4026-91BB-403F90CF8526}.Debug|x64.ActiveCfg = Debug|x64
-		{8B051595-03CD-4026-91BB-403F90CF8526}.Debug|x64.Build.0 = Debug|x64
-		{8B051595-03CD-4026-91BB-403F90CF8526}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8B051595-03CD-4026-91BB-403F90CF8526}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8B051595-03CD-4026-91BB-403F90CF8526}.Release|x64.ActiveCfg = Release|x64
-		{8B051595-03CD-4026-91BB-403F90CF8526}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1DF01FBE-19B3-401E-8A69-EAE343E0AF38}
-	EndGlobalSection
-EndGlobal
