--- samples/csharp/end-to-end-apps/SpikeDetection-ShampooSales-WinForms/ShampooSalesWinforms/Data/shampoo-sales.csv
@@ -1,37 +1,0 @@
-Month,Shampoo sales
-1-Jan,266
-1-Feb,145.9
-1-Mar,183.1
-1-Apr,119.3
-1-May,180.3
-1-Jun,168.5
-1-Jul,231.8
-1-Aug,224.5
-1-Sep,192.8
-1-Oct,122.9
-1-Nov,336.5
-1-Dec,185.9
-2-Jan,194.3
-2-Feb,149.5
-2-Mar,210.1
-2-Apr,273.3
-2-May,191.4
-2-Jun,287
-2-Jul,226
-2-Aug,303.6
-2-Sep,289.9
-2-Oct,421.6
-2-Nov,264.5
-2-Dec,342.3
-3-Jan,339.7
-3-Feb,440.4
-3-Mar,315.9
-3-Apr,439.3
-3-May,401.3
-3-Jun,437.4
-3-Jul,575.5
-3-Aug,407.6
-3-Sep,682
-3-Oct,475.3
-3-Nov,581.3
-3-Dec,646.9
