--- samples/csharp/end-to-end-apps/SpikeDetection-ShampooSales-WinForms/ShampooSalesWinforms/Program.cs
@@ -1,23 +1,0 @@
-using System;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace ShampooSalesAnomalyDetection
-{
-    static class Program
-    {
-        /// <summary>
-        /// The main entry point for the application.
-        /// </summary>
-        [STAThread]
-        static void Main()
-        {
-            Application.EnableVisualStyles();
-            Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
-        }
-    }
-}
