+++ samples/csharp/end-to-end-apps/TimeSeries-ProductSalesSpikeDetection/ProductSalesSpikeDetection/DataStructures/ProductSalesData.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.ML.Data;
+
+namespace ProductSalesSpikeDetection
+{
+    public class ProductSalesData
+    {
+        [LoadColumn(0)]
+        public string Month;
+
+        [LoadColumn(1)]
+        public float numSales;
+    }
+}
