+++ samples/csharp/end-to-end-apps/TimeSeries-ProductSalesSpikeDetection/ProductSalesSpikeDetection/Program.cs
@@ -1,0 +1,23 @@
+using System;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace ProductSalesSpikeDetection
+{
+    static class Program
+    {
+        /// <summary>
+        /// The main entry point for the application.
+        /// </summary>
+        [STAThread]
+        static void Main()
+        {
+            Application.EnableVisualStyles();
+            Application.SetCompatibleTextRenderingDefault(false);
+            Application.Run(new Form1());
+        }
+    }
+}
