+++ samples/csharp/end-to-end-apps/TimeSeries-ProductSalesSpikeDetection/ProductSalesSpikeDetection/Properties/Resources.resx
