+++ samples/csharp/end-to-end-apps/TimeSeries-ProductSalesSpikeDetection/ProductSalesSpikeDetectionE2EApp.sln
@@ -1,0 +1,41 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.539
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProductSalesSpikeDetection", "ProductSalesSpikeDetection\ProductSalesSpikeDetection.csproj", "{8B051595-03CD-4026-91BB-403F90CF8526}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProductSalesSpikeDetectionTrainer", "ProductSalesSpikeDetectionTrainer\ProductSalesSpikeDetectionTrainer.csproj", "{CBCFA004-184B-4EBA-BE50-7B17B3F0E0D6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8B051595-03CD-4026-91BB-403F90CF8526}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8B051595-03CD-4026-91BB-403F90CF8526}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8B051595-03CD-4026-91BB-403F90CF8526}.Debug|x64.ActiveCfg = Debug|x64
+		{8B051595-03CD-4026-91BB-403F90CF8526}.Debug|x64.Build.0 = Debug|x64
+		{8B051595-03CD-4026-91BB-403F90CF8526}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8B051595-03CD-4026-91BB-403F90CF8526}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8B051595-03CD-4026-91BB-403F90CF8526}.Release|x64.ActiveCfg = Release|x64
+		{8B051595-03CD-4026-91BB-403F90CF8526}.Release|x64.Build.0 = Release|x64
+		{CBCFA004-184B-4EBA-BE50-7B17B3F0E0D6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CBCFA004-184B-4EBA-BE50-7B17B3F0E0D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CBCFA004-184B-4EBA-BE50-7B17B3F0E0D6}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{CBCFA004-184B-4EBA-BE50-7B17B3F0E0D6}.Debug|x64.Build.0 = Debug|Any CPU
+		{CBCFA004-184B-4EBA-BE50-7B17B3F0E0D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CBCFA004-184B-4EBA-BE50-7B17B3F0E0D6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CBCFA004-184B-4EBA-BE50-7B17B3F0E0D6}.Release|x64.ActiveCfg = Release|Any CPU
+		{CBCFA004-184B-4EBA-BE50-7B17B3F0E0D6}.Release|x64.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6EFEA7B9-262A-4312-A559-DC88A56734D2}
+	EndGlobalSection
+EndGlobal
