+++ samples/csharp/end-to-end-apps/TimeSeries-ProductSalesSpikeDetection/ProductSalesSpikeDetectionTrainer/Data/SHAMPOO-SALES-LICENSE.txt
@@ -1,0 +1,11 @@
+Product Sales Dataset:
+
+Based on the dataset “Shampoo Sales Over a Three Year Period” originally sourced from DataMarket and provided by Time Series Data Library (TSDL), created by Rob Hyndman. 
+
+“Shampoo Sales Over a Three Year Period” Dataset Licensed Under the DataMarket Default Open License:
+
+License
+You are allowed to copy and redistribute the data as long as you clearly indicate the data provider and DataMarket as the original source.
+
+License summary
+You may copy and redistribute the data. You may make derivative works from the data. You may use the data for commercial purposes. You may not sublicence the data when redistributing it. You may not redistribute the data under a different license. Source attribution on any use of this data: Must refer source.
