+++ samples/csharp/end-to-end-apps/TimeSeries-ProductSalesSpikeDetection/ProductSalesSpikeDetectionTrainer/DataStructures/ProductSalesPrediction.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.ML.Data;
+
+namespace ProductSalesSpikeDetectionTrainer
+{
+    class ProductSalesPrediction
+    {
+        // Vector to hold Alert, Score, and P-Value values
+        [VectorType(3)]
+        public double[] Prediction { get; set; }
+    }
+}
