+++ samples/csharp/end-to-end-apps/TimeSeries-ProductSalesSpikeDetection/ProductSalesSpikeDetectionTrainer/ProductSalesSpikeDetectionTrainer.csproj
