--- samples/csharp/getting-started/SpikeDetection_ShampooSales/ShampooSales.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.438
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ShampooSales", "ShampooSales\ShampooSalesConsoleApp\ShampooSales.csproj", "{95A1AF63-BA25-4F58-BE82-711B3CDA2308}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{95A1AF63-BA25-4F58-BE82-711B3CDA2308}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{95A1AF63-BA25-4F58-BE82-711B3CDA2308}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{95A1AF63-BA25-4F58-BE82-711B3CDA2308}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{95A1AF63-BA25-4F58-BE82-711B3CDA2308}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {04829BD9-3AFA-47D0-8B99-28B41845F774}
-	EndGlobalSection
-EndGlobal
