--- samples/csharp/getting-started/SpikeDetection_ShampooSales/ShampooSales/Data/shampoo-sales.csv
@@ -1,37 +1,0 @@
-Month,Sales of shampoo over a three year period
-1-Jan,266
-2-Jan,145.9
-3-Jan,183.1
-4-Jan,119.3
-5-Jan,180.3
-6-Jan,168.5
-7-Jan,231.8
-8-Jan,224.5
-9-Jan,192.8
-10-Jan,122.9
-11-Jan,336.5
-12-Jan,185.9
-1-Feb,194.3
-2-Feb,149.5
-3-Feb,210.1
-4-Feb,273.3
-5-Feb,191.4
-6-Feb,287
-7-Feb,226
-8-Feb,303.6
-9-Feb,289.9
-10-Feb,421.6
-11-Feb,264.5
-12-Feb,342.3
-1-Mar,339.7
-2-Mar,440.4
-3-Mar,315.9
-4-Mar,439.3
-5-Mar,401.3
-6-Mar,437.4
-7-Mar,575.5
-8-Mar,407.6
-9-Mar,682
-10-Mar,475.3
-11-Mar,581.3
-12-Mar,646.9
