--- samples/csharp/getting-started/SpikeDetection_ShampooSales/ShampooSales/ShampooSalesConsoleApp/DataStructures/ShampooSalesData.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.ML.Data;
-
-namespace ShampooSales.DataStructures
-{
-    public class ShampooSalesData
-    {
-        [LoadColumn(0)]
-        public string Month;
-
-        [LoadColumn(1)]
-        public float numSales;
-    }
-}
