--- samples/csharp/getting-started/SpikeDetection_ShampooSales/ShampooSales/ShampooSalesConsoleApp/ShampooSales.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.ML" Version="$(MicrosoftMLVersion)" />
-    <PackageReference Include="Microsoft.ML.TimeSeries" Version="0.12.0-preview" />
-  </ItemGroup>
-
-</Project>
