+++ samples/csharp/getting-started/TimeSeries_ProductSalesSpikeDetection/ProductSalesSpikeDetection.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.438
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProductSalesSpikeDetection", "ProductSalesSpikeDetection\ProductSalesConsoleApp\ProductSalesSpikeDetection.csproj", "{8F80D65C-B34C-47C6-BBA0-C7B353C32F7D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8F80D65C-B34C-47C6-BBA0-C7B353C32F7D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8F80D65C-B34C-47C6-BBA0-C7B353C32F7D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8F80D65C-B34C-47C6-BBA0-C7B353C32F7D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8F80D65C-B34C-47C6-BBA0-C7B353C32F7D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {04829BD9-3AFA-47D0-8B99-28B41845F774}
+	EndGlobalSection
+EndGlobal
