+++ samples/csharp/getting-started/TimeSeries_ProductSalesSpikeDetection/ProductSalesSpikeDetection/Data/product-sales.csv
@@ -1,0 +1,37 @@
+Month,ProductSales
+1-Jan,271
+2-Jan,150.9
+3-Jan,188.1
+4-Jan,124.3
+5-Jan,185.3
+6-Jan,173.5
+7-Jan,236.8
+8-Jan,229.5
+9-Jan,197.8
+10-Jan,127.9
+11-Jan,341.5
+12-Jan,190.9
+1-Feb,199.3
+2-Feb,154.5
+3-Feb,215.1
+4-Feb,278.3
+5-Feb,196.4
+6-Feb,292
+7-Feb,231
+8-Feb,308.6
+9-Feb,294.9
+10-Feb,426.6
+11-Feb,269.5
+12-Feb,347.3
+1-Mar,344.7
+2-Mar,445.4
+3-Mar,320.9
+4-Mar,444.3
+5-Mar,406.3
+6-Mar,442.4
+7-Mar,580.5
+8-Mar,412.6
+9-Mar,687
+10-Mar,480.3
+11-Mar,586.3
+12-Mar,651.9
