+++ samples/csharp/getting-started/TimeSeries_ProductSalesSpikeDetection/ProductSalesSpikeDetection/ProductSalesConsoleApp/DataStructures/ProductSalesData.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.ML.Data;
+
+namespace ProductSalesSpikeDetection.DataStructures
+{
+    public class ProductSalesData
+    {
+        [LoadColumn(0)]
+        public string Month;
+
+        [LoadColumn(1)]
+        public float numSales;
+    }
+}
