+++ samples/csharp/getting-started/TimeSeries_ProductSalesSpikeDetection/ProductSalesSpikeDetection/ProductSalesConsoleApp/DataStructures/ProductSalesPrediction.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.ML.Data;
+
+namespace ProductSalesSpikeDetection.DataStructures
+{
+    public class ProductSalesPrediction
+    {
+        //vector to hold alert,score,p-value values
+        [VectorType(3)]
+        public double[] Prediction { get; set; }
+    }
+}
