+++ samples/csharp/getting-started/TimeSeries_ProductSalesSpikeDetection/docs/images/ChangePointDetection.png
